attempt 101 for T2

+++ umidov.ali/T2/item_data.h
@@ -1,11 +1,13 @@
 #ifndef ITEM_DATA_H
 #define ITEM_DATA_H
+#include <vector>
+#include <algorithm>
 #include <iostream>
 #include <string>
 #include <sstream>
 namespace myspace {
     struct ItemData {
-        double key1 = 0.0;
+        unsigned long long key1 = 0;
         char key2 = '\0';
         std::string key3;
         friend std::istream& operator>>(std::istream& in, ItemData& item);
